feat: aot virtual calls

Created-by: Abdurakhmanov
Commit-by: Abdurakhmanov
Merged-by: Abdurakhmanov
Description: - kill ectype ireg marks
- aot virtual call support: rewrite & execute
- update asm-tests to format

See merge request: wellox/cbc-engine-initial-stage!54

+++ src/engine/symlevel/aot_table.h
@@ -62,15 +62,15 @@ public:
         Engine::Session& session, IO::FileId fileId, Offset<VirtualCallAotData> offset
     );
 
-    uint32_t GetVNum() const { return vnum; }
+    uint16_t GetVNum() const { return vnum; }
 
-    uint32_t GetExtDefNum() const { return extDefNum; }
+    uint16_t GetExtDefNum() const { return extDefNum; }
 
 private:
-    VirtualCallAotData(uint32_t vnum, uint32_t extDefNum) : vnum(vnum), extDefNum(extDefNum) {}
+    VirtualCallAotData(uint16_t vnum, uint16_t extDefNum) : vnum(vnum), extDefNum(extDefNum) {}
 
-    uint32_t vnum;
-    uint32_t extDefNum;
+    uint16_t vnum;
+    uint16_t extDefNum;
 };
 
 class VirtualCallAotTable {
